add rds-s3 directory in utils and deletion of iam/secrets-manager
