+++ tests/e2e/utils/rds-s3/auto-rds-s3-setup.py
@@ -6,7 +6,8 @@ import yaml
 
 from importlib_metadata import metadata
 from e2e.fixtures.cluster import create_iam_service_account
-from utils import (
+from e2e.utils.config import configure_env_file
+from e2e.utils.utils import (
     get_ec2_client,
     get_rds_client,
     get_eks_client,
@@ -15,15 +16,14 @@ from utils import (
     kubectl_apply,
     print_banner,
     write_yaml_file,
+    load_yaml_file,
+    wait_for,
 )
 
 from shutil import which
-from time import sleep
 
 
 def main():
-    header()
-
     verify_prerequisites()
     s3_client = get_s3_client(
         region=CLUSTER_REGION,
@@ -35,9 +35,9 @@ def main():
     ec2_client = get_ec2_client(CLUSTER_REGION)
     setup_rds(rds_client, secrets_manager_client, eks_client, ec2_client)
     setup_cluster_secrets()
-    setup_kubeflow()
+    setup_kubeflow_pipeline()
 
-    footer()
+    print_banner("RDS S3 Setup Complete")
     script_metadata = [
         f"bucket_name={S3_BUCKET_NAME}",
         f"db_instance_name={DB_INSTANCE_NAME}",
@@ -53,12 +53,8 @@ def main():
         "subnetGroupName": DB_SUBNET_GROUP_NAME,
     }
     write_yaml_file(
-        yaml_content=script_metadata, file_path="utils/auto-rds-s3-setup-metadata.yaml"
+        yaml_content=script_metadata, file_path="utils/rds-s3/metadata.yaml"
     )
-
-
-def header():
-    print_banner("RDS S3 Setup")
 
 
 def verify_prerequisites():
@@ -220,6 +216,7 @@ def get_cluster_private_subnet_ids(eks_client, ec2_client):
         "resourcesVpcConfig"
     ]["subnetIds"]
 
+    # TODO handle pagination
     subnets = ec2_client.describe_subnets(SubnetIds=subnet_ids)["Subnets"]
     private_subnets = []
     for subnet in subnets:
@@ -289,26 +286,22 @@ def get_vpc_security_group_id(eks_client):
 
 
 def wait_for_rds_db_instance_to_become_available(rds_client):
-    status = None
 
     print("Waiting for RDS DB instance to become available...")
-    wait_periods = 30
-    period_length = 30
-    for _ in range(wait_periods):
+
+    def callback():
         status = rds_client.describe_db_instances(
             DBInstanceIdentifier=DB_INSTANCE_NAME
         )["DBInstances"][0]["DBInstanceStatus"]
-        if status == "available":
-            print("RDS DB instance is available!")
-            return
         if status == "failed":
             raise Exception(
                 "An unexpected error occurred while waiting for the RDS DB instance to become available!"
             )
+        assert status == "available"
+        if status == "available":
+            print("RDS DB instance is available!")
 
-        sleep(period_length)
-
-    print("RDS DB instance availability timeout")
+    wait_for(callback, 900)
 
 
 def create_rds_secret(secrets_manager_client, rds_secret_name, rds_root_password):
@@ -402,12 +395,6 @@ def install_secrets_store_csi_driver():
     )
 
 
-def setup_kubeflow():
-    print_banner("Kubeflow Setup")
-
-    setup_kubeflow_pipeline()
-
-
 def setup_kubeflow_pipeline():
     print("Setting up Kubeflow Pipeline...")
 
@@ -422,15 +409,13 @@ def setup_kubeflow_pipeline():
     pipeline_rds_secret_provider_class_file = (
         "../../awsconfigs/common/aws-secrets-manager/rds/secret-provider.yaml"
     )
-    pipeline_rds_params_env_lines = get_pipeline_params_env_lines(
-        pipeline_rds_params_env_file
-    )
-    new_pipeline_rds_params_env_lines = get_updated_pipeline_rds_params_env_lines(
-        db_instance_info, pipeline_rds_params_env_lines
-    )
-    edit_pipeline_params_env_file(
-        new_pipeline_rds_params_env_lines, pipeline_rds_params_env_file
-    )
+
+    rds_params = {
+        "dbHost": db_instance_info["Endpoint"]["Address"],
+        "mlmdDb": db_instance_info["DBName"],
+    }
+    edit_pipeline_params_env_file(rds_params, pipeline_rds_params_env_file)
+
     update_secret_provider_class(
         pipeline_rds_secret_provider_class_file, RDS_SECRET_NAME
     )
@@ -439,92 +424,27 @@ def setup_kubeflow_pipeline():
     pipeline_s3_secret_provider_class_file = (
         "../../awsconfigs/common/aws-secrets-manager/s3/secret-provider.yaml"
     )
-    pipeline_s3_params_env_lines = get_pipeline_params_env_lines(
-        pipeline_s3_params_env_file
-    )
-    new_pipeline_s3_params_env_lines = get_updated_pipeline_s3_params_env_lines(
-        pipeline_s3_params_env_lines
-    )
-    edit_pipeline_params_env_file(
-        new_pipeline_s3_params_env_lines, pipeline_s3_params_env_file
-    )
+
+    s3_params = {
+        "bucketName": S3_BUCKET_NAME,
+        "minioServiceRegion": CLUSTER_REGION,
+        "minioServiceHost": "s3.amazonaws.com",
+    }
+    edit_pipeline_params_env_file(s3_params, pipeline_s3_params_env_file)
+
     update_secret_provider_class(pipeline_s3_secret_provider_class_file, S3_SECRET_NAME)
 
     print("Kubeflow pipeline setup done!")
 
 
-def get_pipeline_params_env_lines(pipeline_params_env_file):
-    with open(pipeline_params_env_file, "r") as file:
-        pipeline_params_env_lines = file.readlines()
-    return pipeline_params_env_lines
-
-
-def get_updated_pipeline_rds_params_env_lines(
-    db_instance_info, pipeline_params_env_lines
-):
-    db_host_pattern = "dbHost="
-    db_host_new_line = db_host_pattern + db_instance_info["Endpoint"]["Address"] + "\n"
-
-    db_mlmd_db_pattern = "mlmdDb="
-    db_mlmd_db_new_line = db_mlmd_db_pattern + db_instance_info["DBName"] + "\n"
-
-    new_pipeline_params_env_lines = []
-
-    for line in pipeline_params_env_lines:
-        line = replace_line(line, db_host_pattern, db_host_new_line)
-        line = replace_line(line, db_mlmd_db_pattern, db_mlmd_db_new_line)
-        new_pipeline_params_env_lines.append(line)
-
-    return new_pipeline_params_env_lines
-
-
-def get_updated_pipeline_s3_params_env_lines(pipeline_params_env_lines):
-    bucket_name_pattern = "bucketName="
-    bucket_name_new_line = bucket_name_pattern + S3_BUCKET_NAME + "\n"
-
-    minio_service_region_pattern = "minioServiceRegion="
-    minio_service_region_new_line = minio_service_region_pattern + CLUSTER_REGION + "\n"
-
-    new_pipeline_params_env_lines = []
-
-    for line in pipeline_params_env_lines:
-        line = replace_line(line, bucket_name_pattern, bucket_name_new_line)
-        line = replace_line(
-            line, minio_service_region_pattern, minio_service_region_new_line
-        )
-        new_pipeline_params_env_lines.append(line)
-
-    return new_pipeline_params_env_lines
-
-
-def replace_line(line, pattern, new_line):
-    if line.startswith(pattern):
-        return new_line
-    else:
-        return line
-
-
-def edit_pipeline_params_env_file(
-    new_pipeline_params_env_lines, pipeline_params_env_file
-):
+def edit_pipeline_params_env_file(params_env, pipeline_params_env_file):
     print(f"Editing {pipeline_params_env_file} with appropriate values...")
 
-    with open(pipeline_params_env_file, "w") as file:
-        file.writelines(new_pipeline_params_env_lines)
-
-
-def read_yaml(filename):
-    with open(filename) as file:
-        return yaml.safe_load(file.read())
-
-
-def write_yaml(filename, contents):
-    with open(filename, "w") as file:
-        yaml.dump(contents, file)
+    configure_env_file(pipeline_params_env_file, params_env)
 
 
 def update_secret_provider_class(secret_provider_class_file, secret_name):
-    secret_provider = read_yaml(secret_provider_class_file)
+    secret_provider = load_yaml_file(secret_provider_class_file)
 
     secret_provider_objects = yaml.safe_load(
         secret_provider["spec"]["parameters"]["objects"]
@@ -534,11 +454,7 @@ def update_secret_provider_class(secret_provider_class_file, secret_name):
         secret_provider_objects
     )
 
-    write_yaml(secret_provider_class_file, secret_provider)
-
-
-def footer():
-    print_banner("RDS S3 Setup Complete")
+    write_yaml_file(secret_provider, secret_provider_class_file)
 
 
 parser = argparse.ArgumentParser()
