+++ tests/e2e/utils/rds-s3/metadata.yaml
@@ -1,0 +1,7 @@
+RDS:
+  instanceName: 
+  secretName: 
+  subnetGroupName:
+S3:
+  bucket: 
+  secretName:
